requirements.txt
certifi==2019.6.16
chardet==3.0.4
future==0.17.1
huepy==0.9.8.1
idna==2.8
instabot==0.4.5
Pillow==6.1.0
python-dotenv==0.10.3
requests==2.22.0
requests-toolbelt==0.9.1
schedule==0.6.0
six==1.12.0
tqdm==4.34.0
urllib3==1.25.3
